+++ hide_menu_user/models/res_user.py
@@ -26,26 +26,26 @@ from odoo import models, fields, api
 class HideMenuUser(models.Model):
     _inherit = 'res.users'
 
-    @api.model
-    def create(self, vals):
+    @api.model_create_multi
+    def create(self, vals_list):
         """
         Else the menu will be still hidden even after removing from the list
         """
         self.clear_caches()
-        return super(HideMenuUser, self).create(vals)
+        return super(HideMenuUser, self).create(vals_list)
 
     def write(self, vals):
         """
         Else the menu will be still hidden even after removing from the list
         """
         res = super(HideMenuUser, self).write(vals)
-        for menu in self.hide_menu_ids:
-            menu.write({
-                'restrict_user_ids': [(4, self.id)]
-            })
+        for record in self:
+            for menu in record.hide_menu_ids:
+                menu.write({
+                    'restrict_user_ids': [(4, record.id)]
+                })
         self.clear_caches()
         return res
-
     def _get_is_admin(self):
         """
         The Hide specific menu tab will be hidden for the Admin user form.
@@ -59,7 +59,7 @@ class HideMenuUser(models.Model):
     hide_menu_ids = fields.Many2many('ir.ui.menu', string="Menu", store=True,
                                      help='Select menu items that needs to be '
                                           'hidden to this user ')
-    is_admin = fields.Boolean(compute=_get_is_admin)
+    is_admin = fields.Boolean(compute=_get_is_admin, string="Admin")
 
 
 class RestrictMenu(models.Model):