+++ product_visibility_website/__manifest__.py
@@ -22,7 +22,7 @@
 ################################################################################
 {
     'name': 'Website Product Visibility',
-    'version': '16.0.1.1.0',
+    'version': '16.0.1.1.2',
     'category': 'Website',
     'summary': 'Website Product visibility for Users',
     'description': """Website Product visibility for Users""",
@@ -33,6 +33,7 @@
     'depends': ['contacts', 'website_sale','website'],
     'data': [
         'views/website_product_visibility.xml',
+        'views/res_config_settings_views.xml'
     ],
     'images': ['static/description/banner.png'],
     'license': 'AGPL-3',