+++ website_hide_variants/__manifest__.py
@@ -21,7 +21,7 @@
 #############################################################################
 {
     'name': "Disable Variants in Website",
-    'version': '16.0.1.0.0',
+    'version': '16.0.2.0.0',
     'description': "The module helps to disable products amd product variants from website",
     'summary': "Hide Variants",
     'author': "Cybrosys Techno Solutions",
@@ -35,7 +35,8 @@
         'website_sale',
         'sale_management'
     ],
-    'data': ['views/product_views.xml'],
+    'data': ['views/product_views.xml',
+             ],
 
     'assets': {
         'web.assets_frontend': [