+++ product_visibility_website/doc/RELEASE_NOTES.md
@@ -1,4 +1,4 @@
-## Module <website_product_visibility>
+## Module <product_visibility_website>
 
 #### 18.08.2022
 #### Version 16.0.1.0.0
@@ -8,3 +8,13 @@
 #### Version 16.0.1.0.0
 ##### FIX
 - Bug Fixed 
+
+#### 20.10.2023
+#### Version 16.0.1.1.1
+##### FIX
+- Feature Of Website Product Visibility For Portal Users Is Added
+
+#### 17.10.2023
+#### Version 16.0.1.1.2
+##### FIX
+- The issue with the product/product-category search functionality has been fixed